remove duplicate date options
