+++ frontend/src/metabase/query_builder/components/filters/modals/BulkFilterSelect/BulkFilterSelect.tsx
@@ -9,6 +9,7 @@ import Dimension from "metabase-lib/lib/Dimension";
 import { isBoolean } from "metabase/lib/schema_metadata";
 
 import TippyPopoverWithTrigger from "metabase/components/PopoverWithTrigger/TippyPopoverWithTrigger";
+import { DateShortcutOptions } from "metabase/query_builder/components/filters/pickers/DatePicker/DatePickerShortcutOptions";
 
 import {
   SelectFilterButton,
@@ -20,6 +21,7 @@ export interface BulkFilterSelectProps {
   query: StructuredQuery;
   filter?: Filter;
   dimension: Dimension;
+  dateShortcutOptions?: DateShortcutOptions;
   customTrigger?: ({ onClick }: { onClick: () => void }) => JSX.Element;
   handleChange: (newFilter: Filter) => void;
   handleClear: () => void;
@@ -29,6 +31,7 @@ export const BulkFilterSelect = ({
   query,
   filter,
   dimension,
+  dateShortcutOptions,
   customTrigger,
   handleChange,
   handleClear,
@@ -66,6 +69,7 @@ export const BulkFilterSelect = ({
           isNew={filter == null}
           showCustom={false}
           showFieldPicker={false}
+          dateShortcutOptions={dateShortcutOptions}
           onChangeFilter={handleChange}
           onClose={closePopover}
           commitOnBlur
